# avoid using reflect and reduce allocs