```
benchmark                old ns/op     new ns/op     delta
BenchmarkParseFile-4     217023226     199891609     -7.89%

benchmark                old allocs     new allocs     delta
BenchmarkParseFile-4     1179043        955346         -18.97%

benchmark                old bytes     new bytes     delta
BenchmarkParseFile-4     45851108      41793534      -8.85%
```